refactor: move method to only place its used (#166)
